+++ README.md
@@ -5,38 +5,78 @@
 [![Test coverage][codecov-image]][codecov-url]
 [![npm download][download-image]][download-url]
 
-Load and parse varian NMR native format.
+Parse varian NMR native format in JS.
 
 ## Installation
 
-`$ npm i varian-converter`
+`npm i varian-converter filelist-from`
 
+## Example of use in Browser
+
+The user may upload files using an `<input>` element, and hand those to varian-convert. This returns an object representing all the parsed data (see the [result blueprint][#Blueprint]).
+
+The `index.html`
+```html
+<!DOCTYPE html>
+<html>
+<input type='file' webkitdirectory mozdirectory directory placeholder='load files...'/>
+<script src="./bundle.js"></script>
+</html>
+```
+Notice `main.js` or `main.ts` has to be bundled `bundle.js`.
+
+* JS for _zipped_ files: `main.js` 
+```javascript
+import {join} from 'path';
+import {fileListFromZip as fromZip} from 'filelist-from';
+import { convert1D as cv } from 'varian-converter';
+
+const input = document.getElementById('input');
+input.onchange = async function (){
+ const filesList = await fromZip(input.files);
+ console.log(cv(fileList));
+ //do stuff
+ return;
+}
+```
+Or a standard _directory_ `<directory>.fid`:
+```
+import {join} from 'path';
+import { convert1D as cv } from 'varian-converter';
+
+const input = document.getElementById('input');
+input.onchange = async function (){
+ const fileList = input.files;
+ console.log(cv(fileList));
+ return;
+}
+```
 ## NodeJS
 
-```javascript
-import {join} from 'path';
+In Node we have access to the system files through the filesystem module (loaded in `fromPath`). Instead of the event system we use `readFileSync`:
 
-import { convert1DFromPath as cvFromPath } from 'varian-converter';
-
-const result = cvFromPath(join(__dirname, "path/to/dir.fid")) 
 ```
-
-Or from zip:
-```javascript
+import {join} from 'path';
 import {readFileSync} from 'fs';
-import {join} from 'path';
+import {fileListFromZip as fromZip} from 'filelist-from';
+import { convert1D as cv } from 'varian-converter';
 
-import { convert1DFromZip as cvFromZip } from 'varian-converter';
-
-const arrayBuffer= readFileSync(join(__dirname, "path/to/dir.fid.zip"))
-
-cvFromZip(arrayBuffer).then(r=>console.log(r))
+const zipBuffer = readFileSync("path/to/file.zip");
+fromZip(zipBuffer).then(r=>console.log(cv(fileList)))
 ```
 
-## Browser
-In the browser, just getting the file as File and passing it to convert1DFromZip should do. Same as
-in node it expects an ArrayBuffer, which is what the File constructor returns.
+If we prefer to load the `<directory>.fid` (no compression), we can also do it:
 
+```javascript
+import {join} from 'path';
+import {fileListFromPath as fromPath} from 'filelist-from';
+import { convert1D as cv } from 'varian-converter';
+
+const fileList = fromPath("path/to/dir.fid");
+const result = cv(fileList);
+```
+
+## Blueprint
 ## License
 
 [MIT](./LICENSE)
