+++ data/gDQCOSY.zip
