+++ package.json
@@ -46,7 +46,7 @@
     "typescript": "^4.6.2"
   },
   "dependencies": {
-    "filelist-from": "^0.2.0",
+    "filelist-from": "^0.2.1",
     "iobuffer": "^5.1.0",
     "ml-spectra-processing": "^11.0.0"
   }
