+++ src/index.ts
@@ -1,6 +1,4 @@
 export * from './convert1D';
-export * from './convert1DFromDir';
-export * from './convert1DFromZip';
 export * from './readBlock';
 export * from './readProcPar';
 export * from './readFileHeader';
